Initial commit
This is a complete implementation of an LMC emulator. In the future, I plan to implement an assembler and disassembler to provide a complete package. Perhaps even a nice GUI, who knows? :p

Currently this own works with the hard-coded program - that will need to be addressed. I will resolve this having cleaned up some of the code.